Applied greiman's proposed change from https://github.com/greiman/SSD1306Ascii/issues/83.
I can't use the above fix since that will break use of String on Arduino boards. Teensy has a different version of String and gcc.

I will modify bool tickerText(TickerState* state, const String &str); So it will work on Teensy and Arduino.

The function was submitted as a pull by a user and I failed to notice that the check on str is not needed since null can not be passed by reference.

The difference between pass-by-reference and pass-by-pointer is that pointers can be NULL or reassigned whereas references cannot. Use pass-by-pointer if NULL is a valid parameter value or if you want to reassign the pointer. Otherwise, use constant or non-constant references to pass arguments.

This is my planed change:

  bool tickerText(TickerState* state, const String &str) {
    return tickerText(state, str.c_str());
  }